requirements.txt
numpy~=1.22.1
matplotlib~=3.5.1
pandas~=1.4.0
tqdm~=4.62.3
sklearn~=0.0
scikit-learn~=1.0.2
scipy~=1.7.3
scikit-image~=0.19.1
tensorflow-gpu~=2.7.0
download~=0.3.5
requests~=2.27.1